block: blk-crypto-fallback: handle data unit split across multiple bvecs
Till now, the blk-crypto-fallback required each crypto data unit to be
contained within a single bvec. It also required the starting offset
of each bvec to be aligned to the data unit size. This patch removes
both restrictions, so that blk-crypto-fallback can handle crypto data
units split across multiple bvecs. blk-crypto-fallback now only requires
that the total size of the bio be aligned to the crypto data unit size.
The buffer that is being read/written to no longer needs to be data unit
size aligned.

This is useful for making the alignment requirements for direct I/O on
encrypted files similar to those for direct I/O on unencrypted files.

Co-developed-by: Eric Biggers <ebiggers@google.com>
Signed-off-by: Eric Biggers <ebiggers@google.com>
Signed-off-by: Satya Tangirala <satyat@google.com>
[lazerl0rd: Backport to Linux 4.19, which requires slightly different
            returns.]
Signed-off-by: Diab Neiroukh <lazerl0rd@thezest.dev>
Signed-off-by: Oktapra Amtono <oktapra.amtono@gmail.com>